FIX: return string-float, cause float is locale-depended (in de 1,55 instead of 1.55)

+++ src/Money/Money.php
@@ -70,10 +70,10 @@ class Money implements \JsonSerializable, Jsonable, Arrayable {
 
 	/**
 	 * amount as decimal with . as decPoint
-	 * @return float
+	 * @return string
 	 */
 	public function normalize() {
-		return (float)bcdiv($this->amount(), $this->currency->unit_factor, $this->currency->decimal_place);
+		return bcdiv($this->amount(), $this->currency->unit_factor, $this->currency->decimal_place);
 	}
 
 	/**
